fix: Fix LuaJIT on Windows
It turned out that with disabling GC64, LuaJIT would immediately
overflow the stack when opening a Lua state. For now, this reverts to
a version from 2019 that works, but I highly doubt that LuaJIT woould
have been broken on Windows for years.

I'll have to investigate, or rather trial-and-error further.